requirements.txt
django
djangorestframework
requests
pytest
pytest - django
pytest - cov
